requirement.txt
asgiref==3.5.2
Django==4.1.1
image==1.5.33
Pillow==9.2.0
pipp==0.0.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
